# FirmaMobileMauiApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FirmaWebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Kontrahent": {
      "get": {
        "tags": [
          "Kontrahent"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KontrahenciForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KontrahenciForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KontrahenciForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Kontrahent"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Kontrahent/{id}": {
      "get": {
        "tags": [
          "Kontrahent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/KontrahenciForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Kontrahent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Kontrahent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Kontrahent"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Produkt/{id}": {
      "get": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Produkt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Produkt"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ProduktUbocznyZK": {
      "get": {
        "tags": [
          "ProduktUbocznyZK"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProduktUbocznyZK"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/ProduktUbocznyZK/{id}": {
      "get": {
        "tags": [
          "ProduktUbocznyZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProduktyUboczneZKForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProduktUbocznyZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProduktUbocznyZK"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ProduktUbocznyZK"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Adres": {
        "required": [
          "glowny",
          "kodPoczowy",
          "kraj",
          "miejscowosc",
          "nrDomu",
          "poczta",
          "siedziba",
          "ulica"
        ],
        "type": "object",
        "properties": {
          "idAdresu": {
            "type": "integer",
            "format": "int32"
          },
          "ulica": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "miejscowosc": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "nrDomu": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "nrLokalu": {
            "type": "string",
            "nullable": true
          },
          "kodPoczowy": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "poczta": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "kraj": {
            "maxLength": 10,
            "minLength": 1,
            "type": "string"
          },
          "siedziba": {
            "type": "boolean"
          },
          "glowny": {
            "type": "boolean"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          }
        },
        "additionalProperties": false
      },
...
      "Cena": {
        "required": [
          "waluta",
          "wartosc"
        ],
        "type": "object",
        "properties": {
          "idCeny": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "idTypuCeny": {
            "type": "integer",
            "format": "int32"
          },
          "wartosc": {
            "type": "number",
            "format": "double"
          },
          "waluta": {
            "minLength": 1,
            "type": "string"
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          },
          "typCeny": {
            "$ref": "#/components/schemas/TypCeny"
          }
        },
        "additionalProperties": false
      },
...
      "KontrahenciForView": {
        "type": "object",
        "properties": {
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "akronim": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string",
            "nullable": true
          },
          "nip": {
            "type": "string",
            "nullable": true
          },
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "sposobPlatnosciNazwa": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Kontrahent": {
        "required": [
          "akronim",
          "idKontrahenta",
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "akronim": {
            "minLength": 1,
            "type": "string"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "nip": {
            "type": "string",
            "default": "000-000-00-00",
            "nullable": true
          },
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "sposobPlatnosci": {
            "$ref": "#/components/schemas/SposobPlatnosci"
          },
          "adres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adres"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PozycjaZZ": {
        "required": [
          "ilosc",
          "rabat"
        ],
        "type": "object",
        "properties": {
          "idPozycjiZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaZakupu": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "rabat": {
            "type": "number",
            "format": "double"
          },
          "dataRealizacji": {
            "type": "string",
            "format": "date-time"
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "zlecenieZakupu": {
            "$ref": "#/components/schemas/ZlecenieZakupu"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
...
      "Produkt": {
        "required": [
          "jednostkaMiary",
          "kod",
          "nazwa",
          "producent",
          "stawkaVatSprzedazy",
          "stawkaVatZakupu"
        ],
        "type": "object",
        "properties": {
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "kod": {
            "minLength": 1,
            "type": "string"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "dodatkowaNazwa": {
            "type": "string",
            "nullable": true
          },
          "jednostkaMiary": {
            "minLength": 1,
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "swW_PKWiU": {
            "type": "string",
            "nullable": true
          },
          "producent": {
            "minLength": 1,
            "type": "string"
          },
          "idTypu": {
            "type": "integer",
            "format": "int32"
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "stawkaVatZakupu": {
            "type": "number",
            "format": "double"
          },
          "stawkaVatSprzedazy": {
            "type": "number",
            "format": "double"
          },
          "odwrotneObciazenie": {
            "type": "boolean"
          },
          "podzielonaPlatnosc": {
            "type": "boolean"
          },
          "stawkaCla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stwkaAkcyzy": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "kodPcn": {
            "type": "string",
            "nullable": true
          },
          "kraj": {
            "type": "string",
            "nullable": true
          },
          "czasKompletacji": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "idDostawcy": {
            "type": "integer",
            "format": "int32"
          },
          "typProduktu": {
            "$ref": "#/components/schemas/TypProduktu"
          },
          "dostawca": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "pozycjaZZ": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PozycjaZZ"
            },
            "nullable": true
          },
          "cena": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cena"
            },
            "nullable": true
          },
          "zlecenieKompletacji": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieKompletacji"
            },
            "nullable": true
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProduktUbocznyZK": {
        "required": [
          "ilosc"
        ],
        "type": "object",
        "properties": {
          "idPrUbocznego": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "zlecenieKompletacji": {
            "$ref": "#/components/schemas/ZlecenieKompletacji"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          }
        },
        "additionalProperties": false
      },
      "ProduktyForView": {
        "type": "object",
        "properties": {
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "kod": {
            "type": "string",
            "nullable": true
          },
          "nazwa": {
            "type": "string",
            "nullable": true
          },
          "dodatkowaNazwa": {
            "type": "string",
            "nullable": true
          },
          "jednostkaMiary": {
            "type": "string",
            "nullable": true
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "swW_PKWiU": {
            "type": "string",
            "nullable": true
          },
          "producent": {
            "type": "string",
            "nullable": true
          },
          "idTypu": {
            "type": "integer",
            "format": "int32"
          },
          "typProduktuNazwa": {
            "type": "string",
            "nullable": true
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "stawkaVatZakupu": {
            "type": "number",
            "format": "double"
          },
          "stawkaVatSprzedazy": {
            "type": "number",
            "format": "double"
          },
          "odwrotneObciazenie": {
            "type": "boolean"
          },
          "podzielonaPlatnosc": {
            "type": "boolean"
          },
          "stawkaCla": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stwkaAkcyzy": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "kodPcn": {
            "type": "string",
            "nullable": true
          },
          "kraj": {
            "type": "string",
            "nullable": true
          },
          "czasKompletacji": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "czyAktywna": {
            "type": "boolean"
          },
          "idDostawcy": {
            "type": "integer",
            "format": "int32"
          },
          "dostawcaDane": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProduktyUboczneZKForView": {
        "type": "object",
        "properties": {
          "idPrUbocznego": {
            "type": "integer",
            "format": "int32"
          },
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "zlecenieKompletacjiDane": {
            "type": "string",
            "nullable": true
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "produktDane": {
            "type": "string",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "czyAktywny": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "SposobPlatnosci": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idSposobuPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "kontrahent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kontrahent"
            },
            "nullable": true
          },
          "zlecenieZakupu": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ZlecenieZakupu"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TypProduktu": {
        "required": [
          "nazwa"
        ],
        "type": "object",
        "properties": {
          "idTypuProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "minLength": 1,
            "type": "string"
          },
          "uwagi": {
            "type": "string",
            "nullable": true
          },
          "fotoURL": {
            "type": "string",
            "nullable": true
          },
          "czySklep": {
            "type": "boolean"
          },
          "czyAktywny": {
            "type": "boolean"
          },
          "produkt": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Produkt"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ZlecenieKompletacji": {
        "required": [
          "dataWystawienia",
          "ilosc",
          "oczekiwanaDataRealizacji"
        ],
        "type": "object",
        "properties": {
          "idZleceniaKompletacji": {
            "type": "integer",
            "format": "int32"
          },
          "idKontrahenta": {
            "type": "integer",
            "format": "int32"
          },
          "dataWystawienia": {
            "type": "string",
            "format": "date-time"
          },
          "oczekiwanaDataRealizacji": {
            "type": "string",
            "format": "date-time"
          },
          "potwierdzonaDataRealizacji": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idProduktu": {
            "type": "integer",
            "format": "int32"
          },
          "ilosc": {
            "type": "number",
            "format": "double"
          },
          "idMontera": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "priorytet": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "czyAktywne": {
            "type": "boolean"
          },
          "czasZlecenia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kontrahent": {
            "$ref": "#/components/schemas/Kontrahent"
          },
          "monter": {
            "$ref": "#/components/schemas/Monter"
          },
          "produkt": {
            "$ref": "#/components/schemas/Produkt"
          },
          "produktUbocznyZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProduktUbocznyZK"
            },
            "nullable": true
          },
          "skladnikZK": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkladnikZK"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}